+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - search grounding with Gemini models now also available to be used with other tools
 - properly using overridden custom token cost values for model
 - show info that Export to Image of messages is not available when using virtualized rendering
+- added support for Web Search tool for OpenAI models
 
 ## [0.34.0]